tests/wpt/tests/paint-timing/first-paint-only/sibling-painting-first-image.html

<!DOCTYPE html>
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../resources/utils.js"></script>
 <!-- This iframe will have a sibling that paints, we want to ensure it does not detect that paint. -->
<iframe id="listening-iframe" src="../resources/subframe-sending-paint.html"></iframe>

<script>
  setup({"hide_test_state": true});

  const entriesExpectToReceive = [
    {
      'entryType': 'paint',
      'name': 'first-paint'
    },
    {
      'entryType': 'paint',
      'name': 'first-contentful-paint'
    }
  ];

  promise_test(async t => {
    assert_implements(window.PerformancePaintTiming, "Paint Timing isn't supported.");

    let result = await new Promise(resolve => {
      let paintingIframeHasDispatchedEntries = false;
      window.addEventListener('message', e => {
        if (!paintingIframeHasDispatchedEntries) {
          // Check paint-timing entries from the painting iframe.
          for (let i = 0; i < entriesExpectToReceive.length; i++) {
            if (entriesExpectToReceive[i].entryType == e.data.entryType &&
              entriesExpectToReceive[i].name == e.data.name) {
              entriesExpectToReceive.splice(i, 1);
              break;
            }
          }
          if (entriesExpectToReceive.length == 0) {
            paintingIframeHasDispatchedEntries = true;
            // Ask the listening iframe to send its paint-timing entries.
            document.getElementById('listening-iframe').
                contentWindow.postMessage('', '*');
          }
          return;
        }
        // Done. The result is checked in the main test body.
        resolve(e.data);
      });
    });

    // Check the paint-timing entries from the listening iframe.
    assert_equals(result, '0');

    // Check that current frame receives first-paint but not first-contentful-paint.
    await assertFirstPaint();
    await assertNoFirstContentfulPaint(t);
  }, 'Frame ignores paint-timing events fired from sibling frame.');
</script>

<!-- This iframe is where all of the painting occurs. -->
<iframe id="painting-iframe" src="../resources/subframe-painting.html"></iframe>
</body>
</html>
